% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_tools.R
\name{add_forecaster}
\alias{add_forecaster}
\title{add a forecaster to the registry as an algorithm}
\usage{
add_forecaster(forecaster_name, forecaster, reg = getDefaultRegistry())
}
\description{
because there is quite a bit of shared code around evaluating a given
forecaster, we don't expect the user to define the slide and other
evaluation details.
}
