% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_utils.R
\name{add_id}
\alias{add_id}
\title{add a unique id based on the column contents}
\usage{
add_id(df, name_length = 5)
}
\description{
feed a character represenation of the column contents through md5 and reencoding in base64 to get short unique hashes
make sure that there are no columns with \code{NA}'s
}
