% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{add_quantiles}
\alias{add_quantiles}
\title{make sure to have a quantile forecast}
\usage{
add_quantiles(frost, trainer, args_list)
}
\arguments{
\item{frost}{an \code{\link[epipredict:frosting]{epipredict::frosting}}}

\item{args_list}{an \code{\link[epipredict:arx_args_list]{epipredict::arx_args_list}}}
}
\description{
dispatch on the trainer to make sure that the right kind of quantiles are added
}
\seealso{
\link{arx_preprocess} for the step equivalent
}
