% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{arx_basics}
\alias{arx_basics}
\title{predict, na omit, threshold and add dates}
\usage{
arx_basics(frost, forecast_date = NULL, target_date = NULL, nonneg = TRUE)
}
\arguments{
\item{frost}{an \code{\link[epipredict:frosting]{epipredict::frosting}}}

\item{forecast_date}{the date from which the forecast was made. defaults to
the default of \code{layer_add_forecast_date}, which is currently the max
time_value present in the data}

\item{target_date}{the date about which the forecast was made. defaults to
the default of \code{layer_add_target_date}, which is either
\code{forecast_date+ahead}, or the \verb{max time_value + ahead}}
}
\description{
add some basic layers that make sure a prediction is present, that the values
in the prediction are thresholded to be positive, and that the default
forecast date and target date are available as columns.
}
