% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipredict_utilities.R
\name{arx_postprocess}
\alias{arx_postprocess}
\title{add the default layers for arx_forecaster}
\usage{
arx_postprocess(
  postproc,
  trainer,
  args_list,
  forecast_date = NULL,
  target_date = NULL
)
}
\arguments{
\item{postproc}{an \code{\link[epipredict:frosting]{epipredict::frosting}}}

\item{trainer}{the trainer used (e.g. linear_reg() or quantile_reg())}

\item{args_list}{an \code{\link[epipredict:arx_args_list]{epipredict::arx_args_list}}}

\item{forecast_date}{the date from which the forecast was made. defaults to
the default of \code{layer_add_forecast_date}, which is currently the max
time_value present in the data}

\item{target_date}{the date about which the forecast was made. defaults to
the default of \code{layer_add_target_date}, which is either
\code{forecast_date+ahead}, or the \verb{max time_value + ahead}}
}
\description{
add the default layers for arx_forecaster
}
\seealso{
\link{arx_preprocess} for the step equivalent
}
