% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipredict_utilities.R
\name{arx_preprocess}
\alias{arx_preprocess}
\title{add the default steps for arx_forecaster}
\usage{
arx_preprocess(preproc, outcome, predictors, args_list)
}
\arguments{
\item{preproc}{an \code{\link[epipredict:epi_recipe]{epipredict::epi_recipe}}}

\item{outcome}{a character of the column to be predicted}

\item{predictors}{a character vector of the columns used as predictors}

\item{args_list}{an \code{\link[epipredict:default_args_list]{epipredict::default_args_list}}}
}
\description{
add the default steps for arx_forecaster
}
\seealso{
\link{arx_postprocess} for the layer equivalent
}
