% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_lastminute_nas}
\alias{clear_lastminute_nas}
\title{temporary patch that pulls \code{NA}'s out of an epi_df}
\usage{
clear_lastminute_nas(epi_data, outcome, extra_sources)
}
\arguments{
\item{epi_data}{the epi_df to be fixed}
}
\description{
just delete rows that have NA's in them. eventually epipredict should directly handle this so we don't have to
}
