% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{confirm_insufficient_data}
\alias{confirm_insufficient_data}
\title{confirm that there's enough data to run this model}
\usage{
confirm_insufficient_data(epi_data, ahead, args_input, buffer = 9)
}
\arguments{
\item{buffer}{how many training data to insist on having (e.g. if \code{buffer=1},
this trains on one sample; the default is set so that \code{linear_reg} isn't
rank deficient)}
}
\description{
epipredict is a little bit fragile about having enough data to train; we want
to be able to return a null result rather than error out; this check say to
return a null
}
