% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{covidhub_probs}
\alias{covidhub_probs}
\title{the quantile levels used by the covidhub repository}
\usage{
covidhub_probs(type = c("standard", "inc_case"))
}
\arguments{
\item{type}{either standard or inc_case, with inc_case being a small subset of the standard}
}
\description{
the quantile levels used by the covidhub repository
}
