% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_average.R
\name{ensemble_average}
\alias{ensemble_average}
\title{an ensemble model that averages each quantile separately}
\usage{
ensemble_average(
  epi_data,
  other_forecasts,
  outcome,
  extra_sources = "",
  ensemble_args = list(),
  ensemble_args_names = NULL
)
}
\arguments{
\item{epi_data}{the actual data used}

\item{other_forecasts}{a list of quantile forecasts to aggregate. They should
be tibbles with columns \verb{(geo_value, forecast_date, target_end_date, quantile, value)}, preferably in that order.}

\item{outcome}{the name of the target variable}

\item{extra_sources}{the name of any extra columns to use. This list could be
empty}

\item{ensemble_args}{any arguments unique to this particular ensembler should
be included in a list like this (unfortunate targets issues). The arguments
for \code{ensemble_average} in particular are \code{average_type} and \code{join_columns}}

\item{ensemble_args_names}{an argument purely for use in targets. You
probably shouldn't worry about it. In a target, it should probably be
\code{ensemble_args_names = names(ensemble_args)}}
}
\description{
The simplest calss of ensembing models, it takes in a list of quantile
forecasts and averages them on a per-quantile basis. By default the average
used is the median, but it can accept any vectorized function.
}
