% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensemble_missing_forecasters_details}
\alias{ensemble_missing_forecasters_details}
\title{given an ensemble and a list of forecasters used in some of those ensembles, return the ones that use them}
\usage{
ensemble_missing_forecasters_details(
  ensemble_grid = NULL,
  param_grid = NULL,
  project = NULL
)
}
\arguments{
\item{ensemble_grid}{the grid of ensembles used}

\item{param_grid}{the grid of forecasters used that we're checking for presence}

\item{project}{the project to be used; by default, the environmental variable is used}
}
\description{
given an ensemble and a list of forecasters used in some of those ensembles, return the ones that use them
}
