% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration-tooling-package.R
\docType{package}
\name{epieval-package}
\alias{epieval}
\alias{epieval-package}
\title{epieval: Evaluating Timeseries Forecasting on Archival Data}
\description{
Given a timeseries and accompanying auxillary timeseries, evaluate a collection of forecasters implementable with epipredict using several metrics on historical data. Assuming the source of your timeseries provides versioned data, the evaluation at any given timepoint will only use data that was available at that point.
}
\author{
\strong{Maintainer}: David Weber \email{davidweb@andrew.cmu.edu} [contributor]

Authors:
\itemize{
  \item Dmitry Shemetov \email{dshemeto@andrew.cmu.edu}
}

}
\keyword{internal}
