% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latency_adjusting.R
\name{extend_ahead}
\alias{extend_ahead}
\title{only extend the ahead}
\usage{
extend_ahead(epi_data, ahead)
}
\arguments{
\item{epi_data}{the dataset}

\item{ahead}{how many units (depending on the dataset, normally days or weeks) to predict ahead of the \code{forecast_date}}
}
\description{
instead of filling in new values, this just extends how far into the future
the model is predicting. For example, if the last data is on the 3rd, the
\code{as_of} is the 5th, and we want an ahead of 4, then this actually sets the
ahead to be 6, since the 9th (the target date) is 6 days after the last day
of data.
}
