% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forecaster_lookup}
\alias{forecaster_lookup}
\title{look up forecasters by name}
\usage{
forecaster_lookup(forecaster_name, param_grid = NULL, project = NULL)
}
\arguments{
\item{forecaster_name}{a part of the adj.adj.1 name used to identify the forecaster.}

\item{param_grid}{the tibble containing the mapping between}

\item{project}{the project to be used; by default, the environmental variable is used}
}
\description{
given a (partial) forecaster name, look up all forecasters in the given project which contain part of that name.
}
