% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{forecaster_pred}
\alias{forecaster_pred}
\title{generate forecaster predictions on a particular dataset}
\usage{
forecaster_pred(
  data,
  job,
  instance,
  forecaster = scaled_pop,
  slide_training = Inf,
  slide_training_pad = 20L,
  trainer = "linear_reg",
  ahead = 1,
  ...
)
}
\arguments{
\item{data}{as per batchtools}

\item{job}{as per batchtools}

\item{instance}{as per batchtools}

\item{forecaster}{a function that does the actual forecasting for a given
day. See \code{exampleSpec.R} for an example function and its documentation for
the general parameter requirements.}

\item{slide_training}{a required parameter that governs the window size that
epix_slide hands off to epipredict.}

\item{slide_training_pad}{a required parameter that determines padding}

\item{trainer}{should be given as a string, which will be converted to a
function.}

\item{ahead}{a necessary parameter to specify an experiment}

\item{...}{any extra parameters the user has defined for forecaster.}

\item{scoring_function}{TODO detailed spec}
}
\description{
a wrapper that turns a forecaster, parameters, data
combination into an actual experiment that outputs a
prediction.
as far as batchtools is concerned, the scoring function is a particular
parameter of the forecaster (or Algorithm, as they call it).
}
