% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{format_covidhub}
\alias{format_covidhub}
\title{format for the COVID-19 Forecast Hub}
\usage{
format_covidhub(pred, true_forecast_date, target_end_date, levels)
}
\arguments{
\item{pred}{the output as produced by epipredict}

\item{true_forecast_date}{the actual date from which the model is
making the forecast, rather than the last day of available data}

\item{target_end_date}{the date of the prediction}

\item{levels}{the quantile levels}
}
\description{
expects the pred to have the columns geo_value, .pred, .pred_distn,
forecast_date, and target_date
It does not assume that the forecast_date is accurate
The end result
}
