% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_exclusions}
\alias{get_exclusions}
\title{Get exclusions from a JSON file for a given date}
\usage{
get_exclusions(
  date,
  exclusions_json = here::here("scripts", "geo_exclusions.json")
)
}
\arguments{
\item{date}{A date}

\item{exclusions_json}{A JSON file with exclusions in the format:

{"exclusions": {"2024-03-24": "ak,hi"}}}
}
\description{
Get exclusions from a JSON file for a given date
}
