% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{get_nonkey_names}
\alias{get_nonkey_names}
\title{just the names which aren't keys for an epi_df}
\usage{
get_nonkey_names(epi_data)
}
\arguments{
\item{epi_data}{the epi_df}
}
\description{
names, but it excludes keys
}
