% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{get_trainable_names}
\alias{get_trainable_names}
\title{extract the non-key, non-smoothed columns from epi_data}
\usage{
get_trainable_names(epi_data, cols)
}
\arguments{
\item{epi_data}{the epi_data tibble}

\item{cols}{vector of column names to use. If \code{NULL}, fill with all non-key columns}
}
\description{
extract the non-key, non-smoothed columns from epi_data
}
\keyword{internal}
