% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_utils.R
\name{id_ahead_ensemble_grid}
\alias{id_ahead_ensemble_grid}
\title{add aheads, forecaster_ids, and ids to a list of ensemble models}
\usage{
id_ahead_ensemble_grid(ensemble_grid, aheads, n_adj = 2)
}
\arguments{
\item{ensemble_grid}{the list of ensembles,}

\item{aheads}{the aheads to add}

\item{n_adj}{the number of adjectives to use; default of 2.}
}
\description{
minor utility
}
