% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latency_adjusting.R
\name{locf_latency}
\alias{locf_latency}
\title{last observation carried forward}
\usage{
locf_latency(epi_data, ahead)
}
\arguments{
\item{epi_data}{the dataset}

\item{ahead}{how many units (depending on the dataset, normally days or weeks) to predict ahead of the \code{forecast_date}}
}
\description{
instead of modifying \code{ahead}, interpolate \code{epi_data} to contain last
observation carried forward
}
