% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_utils.R
\name{make_external_names_and_scores}
\alias{make_external_names_and_scores}
\title{Make external names and scores targets}
\usage{
make_external_names_and_scores()
}
\description{
Make external names and scores targets
}
