% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_utils.R
\name{make_forecasts_and_scores_by_ahead}
\alias{make_forecasts_and_scores_by_ahead}
\title{Make forecasts and scores by ahead targets}
\usage{
make_forecasts_and_scores_by_ahead()
}
\description{
Make forecasts and scores by ahead targets
}
