% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_utils.R
\name{make_shared_ensembles}
\alias{make_shared_ensembles}
\title{Make list of common ensembles for forecasting experiments across projects}
\usage{
make_shared_ensembles()
}
\description{
Make list of common ensembles for forecasting experiments across projects
}
