% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_utils.R
\name{make_shared_grids}
\alias{make_shared_grids}
\title{Make list of common forecasters for forecasting experiments across projects}
\usage{
make_shared_grids()
}
\description{
Make list of common forecasters for forecasting experiments across projects
}
