% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_S3_cache.R
\name{manage_S3_forecast_cache}
\alias{manage_S3_forecast_cache}
\title{Manage S3 cache}
\usage{
manage_S3_forecast_cache(
  rel_cache_dir,
  bucket_name = "forecasting-team-data",
  direction = "download",
  verbose = FALSE
)
}
\arguments{
\item{rel_cache_dir}{The relative path to the cache directory, e.g.
"data-processed/2021-09-01".}

\item{bucket_name}{The name of the S3 bucket to sync.}

\item{direction}{Set 'download' to download files or 'upload' to upload
files.}

\item{verbose}{Set to TRUE to print the files being synced.}
}
\description{
Manage S3 cache
}
