% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{perform_sanity_checks}
\alias{perform_sanity_checks}
\title{helper function for those writing forecasters}
\usage{
perform_sanity_checks(epi_data, outcome, predictors, trainer, args_list)
}
\arguments{
\item{outcome}{the name of the target variable}

\item{trainer}{an example extra parameter that is fairly common}

\item{args_list}{the args list created by \code{\link[epipredict:arx_args_list]{epipredict::arx_args_list}}}
}
\description{
a smorgasbord of checks that any epipredict-based forecaster should do:
\enumerate{
\item check that the args list is created correctly,
\item rewrite an empty extra sources list from an empty string
\item validate the outcome and predictors as present,
\item make sure the trainer is a \code{regression} model from \code{parsnip}
\item adjust the trainer's quantiles based on those in args_list if it's a quantile trainer
\item remake the lags to match the numebr of predictors
}
}
