% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_utils.R
\name{read_external_predictions_data}
\alias{read_external_predictions_data}
\title{import prediction cards generated elsewhere}
\usage{
read_external_predictions_data(predictions_filename, forecaster_name = NULL)
}
\arguments{
\item{predictions_filename}{the filename to be read as an RDS}

\item{forecaster_name}{the name to assign the forecaster}
}
\description{
load an externally generated RDS to be evaluated
}
