% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{rolling_mean}
\alias{rolling_mean}
\title{get a rolling average for the named columns}
\usage{
rolling_mean(epi_data, width = 7L, cols_to_mean = NULL)
}
\arguments{
\item{epi_data}{the dataset}

\item{width}{the number of days (or examples, the sliding isn't time-aware) to use}

\item{cols_to_mean}{the non-key columns to take the mean over. \code{NULL} means all}
}
\description{
add column(s) that are the rolling means of the specified columns, as
implemented by slider. Defaults to the previous 7 days.
Currently only group_by's on the geo_value. Should probably extend to more
keys if you have them
}
