% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{rolling_sd}
\alias{rolling_sd}
\title{get a rolling standard deviation for the named columns}
\usage{
rolling_sd(
  epi_data,
  sd_width = 28L,
  mean_width = NULL,
  cols_to_sd = NULL,
  keep_mean = FALSE
)
}
\arguments{
\item{epi_data}{the dataset}

\item{sd_width}{the number of days (or examples, the sliding isn't
time-aware) to use for the standard deviation calculation}

\item{mean_width}{like \code{sd_width}, but it governs the mean. Should be less
than the \code{sd_width}, and if \code{NULL} (the default) it is half of \code{sd_width}
(so 14 in the complete default case)}

\item{cols_to_sd}{the non-key columns to take the sd over. \code{NULL} means all}

\item{keep_mean}{bool, if \code{TRUE}, it retains keeps the mean column}
}
\description{
A rolling standard deviation, based off of a rolling mean. First it
calculates a rolling mean with width \code{mean_width}, and then squares the
difference between that and the actual value, averaged over \code{sd_width}.
}
