% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_utils.R
\name{run_evaluation_measure}
\alias{run_evaluation_measure}
\title{evaluate_predictions wrapper}
\usage{
run_evaluation_measure(data, evaluation_data, measures)
}
\arguments{
\item{data}{a prediction card to be scored}

\item{evaluation_data}{the true values}

\item{measures}{a set of scores to be used}
}
\description{
run the measures on \code{data}, with truth data \code{evaluation_data}
}
