% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{run_workflow_and_format}
\alias{run_workflow_and_format}
\title{helper function to run a epipredict model and reformat to hub format}
\usage{
run_workflow_and_format(preproc, postproc, trainer, epi_data)
}
\arguments{
\item{preproc}{the preprocessing steps}

\item{postproc}{the postprocessing frosting}

\item{trainer}{the parsnip trainer}

\item{epi_data}{the actual epi_df to train on}
}
\description{
helper function to run a epipredict model and reformat to hub format
}
