% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{scoring_lookup}
\alias{scoring_lookup}
\title{tables don't play nicely with functions directly}
\usage{
scoring_lookup(string_name)
}
\description{
tables don't play nicely with functions directly
}
