% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_tools.R
\name{set_registry}
\alias{set_registry}
\title{setup the batchtools registry}
\usage{
set_registry(registry_folder = here::here("registry"))
}
\arguments{
\item{registry_folder}{where the registry is stored}
}
\value{
\code{\link{NULL}}
}
\description{
settings come from. Can either load an existing registry, or make a new one
in the case it doesn't exist
}
