% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_utils.R
\name{single_id}
\alias{single_id}
\title{generate an id from a simple list of parameters}
\usage{
single_id(param_list, ahead = NULL, n_adj = 2)
}
\arguments{
\item{param_list}{the list of parameters. must include \code{ahead} if \code{ahead = NULL}}

\item{ahead}{the ahead to use.}

\item{n_adj}{the number of adjectives to use; default of 2.}
}
\description{
generate an id from a simple list of parameters
}
