% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster_smoothed_scaled.R
\name{smooth_scaled}
\alias{smooth_scaled}
\title{predict on smoothed data and the standard deviation}
\usage{
smooth_scaled(
  epi_data,
  outcome,
  extra_sources = "",
  ahead = 1,
  pop_scaling = TRUE,
  trainer = parsnip::linear_reg(),
  quantile_levels = covidhub_probs(),
  smooth_days = 7,
  smooth_cols = NULL,
  sd_days = 28,
  sd_cols = NULL,
  ...
)
}
\arguments{
\item{epi_data}{the actual data used}

\item{outcome}{the name of the target variable}

\item{extra_sources}{the name of any extra columns to use. This list could be
empty}

\item{ahead}{(this is relative to the \code{as_of} field of the \code{epi_df}, which is
likely \emph{not} the same as the \code{ahead} used by epipredict, which is relative
to the max time value of the \code{epi_df}. how to handle this is a modelling
question left up to each forecaster; see latency_adjusting.R for the
existing examples)}

\item{pop_scaling}{an example extra parameter unique to this forecaster}

\item{trainer}{an example extra parameter that is fairly common}

\item{quantile_levels}{The quantile levels to predict. Defaults to those
required by covidhub.}

\item{smooth_days}{the number of days over which to do smoothing. If \code{NULL},
then no smoothing is applied.}

\item{smooth_cols}{the names of the columns to smooth. If \code{NULL} it smooths
everything}

\item{sd_cols}{the names of the columns to smooth. If \code{NULL} its includes
the sd ofeverything}

\item{sd_width}{the number of days over which to take a moving average of the
standard deviation. If \code{NULL}, the sd_width isn't included.}
}
\description{
This is a variant of \code{scaled_pop}, which predicts on a smoothed version of
the data. Even if the target is smoothed when used as a /predictor/, as a
/target/ it still uses the raw value (this captures some of the noise).  It
also uses a rolling standard deviation as an auxillary signal, window of
withd \code{sd_width}, which by default is 28 days.
}
\seealso{
some utilities for making forecasters: \link{format_storage},
\link{perform_sanity_checks}
}
