% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_utils.R
\name{sym_subset}
\alias{sym_subset}
\title{function to map}
\usage{
sym_subset(param_list, sym_names = list("average_type"))
}
\arguments{
\item{sym_names}{a list of the parameter names that should be turned into symbols}
}
\description{
function to map
}
\keyword{internal}
