% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{update_predictors}
\alias{update_predictors}
\title{update the predictors to only contain the smoothed/sd versions of cols}
\usage{
update_predictors(epi_data, cols_modified, predictors)
}
\arguments{
\item{epi_data}{the epi_df}

\item{cols}{the list of columns}
}
\description{
should only be applied after both rolling_mean and rolling_sd
}
