% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_id}
\alias{add_id}
\title{add a unique id based on the column contents}
\usage{
add_id(df, n_adj = 2)
}
\arguments{
\item{df}{the df to add a column to. everything should be convertable to a string}

\item{n_adj}{the number of adjectives to use; default of 2.}
}
\description{
create a string of \code{n_adj} that is a hash of the parameters
and append the \code{ahead} at the end.
}
