% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_utils.R
\name{covidhub_probs}
\alias{covidhub_probs}
\title{the quantile levels used by the covidhub repository}
\usage{
covidhub_probs(type = c("standard", "inc_case"))
}
\description{
the quantile levels used by the covidhub repository
}
