% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{forecaster_pred}
\alias{forecaster_pred}
\title{generate forecaster predictions on a particular dataset}
\usage{
forecaster_pred(
  data,
  outcome,
  extra_sources = "",
  forecaster = scaled_pop,
  slide_training = Inf,
  slide_training_pad = 20L,
  n_training = 32,
  n_training_pad = 0,
  forecaster_args = list(),
  forecaster_args_names = list()
)
}
\arguments{
\item{data}{as per batchtools}

\item{forecaster}{a function that does the actual forecasting for a given
day. See \code{exampleSpec.R} for an example function and its documentation for
the general parameter requirements.}

\item{slide_training}{a required parameter that governs the window size that
epix_slide hands off to epipredict. Note that}

\item{slide_training_pad}{a required parameter that determines how much extra
to hand-off to guarantee that at least \code{slide_training} examples are passed
on (e.g. b/c of missing data).}

\item{job}{as per batchtools}

\item{instance}{as per batchtools}

\item{trainer}{should be given as a string, which will be converted to a
function.}

\item{ahead}{a necessary parameter to specify an experiment}

\item{...}{any extra parameters the user has defined for forecaster.}
}
\description{
a wrapper that turns a forecaster, parameters, data
combination into an actual experiment that outputs a
prediction.
as far as batchtools is concerned, the scoring function is a particular
parameter of the forecaster (or Algorithm, as they call it).
}
