% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{format_storage}
\alias{format_storage}
\title{storage format}
\usage{
format_storage(pred, true_forecast_date, target_end_date)
}
\description{
the columns for any stored prediction are
\verb{(geo_value, forecast_date, target_end_date, quantile, value)}
in that particular order. It does not include the point estimate as a
separate row.
}
