% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_utils.R
\name{make_data_targets}
\alias{make_data_targets}
\title{Make common targets for fetching data}
\usage{
make_data_targets()
}
\description{
Relies on the following globals:
\itemize{
\item \code{hhs_signal}
\item \code{chng_signal}
\item \code{fetch_args}
}
}
