% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_utils.R
\name{make_target_param_grid}
\alias{make_target_param_grid}
\title{convert a list of forecasters}
\usage{
make_target_param_grid(
  param_grid,
  actually_symbol = c("forecaster", "trainer")
)
}
\description{
the required format for targets is a little jank; this takes a human legible tibble and makes it targets legible
}
