% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_S3.R
\name{manage_S3_forecast_cache}
\alias{manage_S3_forecast_cache}
\title{Manage S3 cache}
\usage{
manage_S3_forecast_cache(
  rel_cache_dir = NULL,
  bucket_name = "forecasting-team-data",
  direction = "sync",
  verbose = FALSE,
  prefix = Sys.getenv("AWS_S3_PREFIX", "exploration"),
  tar_project = Sys.getenv("TAR_PROJECT", ""),
  external_scores_path = Sys.getenv("EXTERNAL_SCORES_PATH",
    "legacy-exploration-scorecards.qs")
)
}
\arguments{
\item{rel_cache_dir}{The relative path to the cache directory, e.g.
"data-processed/2021-09-01". Default is \code{"{tar_project}/objects"}}

\item{bucket_name}{The name of the S3 bucket to sync.}

\item{direction}{Set 'download' to download files, 'upload' to upload files,
and \code{sync} to do both}

\item{verbose}{Set to TRUE to print the files being synced.}

\item{prefix}{specify the prefix for \code{s3sync}, which filters down which files
to sync to those starting with \code{prefix}.}

\item{tar_project}{which targets project we're working on}

\item{external_scores_path}{the file location where the legacy scores should be saved}
}
\description{
Manage S3 cache
}
